requirements_hgt.txt
torch>=2.0.0
torch-geometric>=2.4.0
transformers>=4.51.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
Pillow>=9.0.0
sentence-transformers>=2.2.0
